Trait rel_alloc::EmplaceIn

pub trait EmplaceIn<A: RegionalAllocator> {
    fn emplace_in<T>(self, alloc: A) -> OwnedVal<T, A>
    where
        T: DropRaw + Pointee + ?Sized,
        <T as Pointee>::Metadata: Metadata<T>,
        Self: Emplace<T, A::Region>;
}

An extension trait for Emplace that provides an allocating emplacement function.

Required Methods

fn emplace_in<T>(self, alloc: A) -> OwnedVal<T, A>where
T: DropRaw + Pointee + ?Sized,
<T as Pointee>::Metadata: Metadata<T>,
Self: Emplace<T, A::Region>,

Emplaces a value into a new OwnedVal allocated from the given allocator and returns it.