Struct rel_alloc::vec::RelVec

#[repr(C)]
pub struct RelVec<T, A: RawRegionalAllocator, B: Basis = DefaultBasis> { /* private fields */ }

A relative counterpart to Vec.

Implementations

impl<T, A: RawRegionalAllocator, B: Basis> RelVec<T, A, B>

pub fn is_empty(&self) -> bool

Returns true if the RelVec contains no elements.

pub fn len(&self) -> usize

Returns the number of elements in the RelVec.

pub fn as_ptr(this: Ref<'_, Self>) -> *const T

Returns a raw pointer to the RelVec’s buffer, or a dangling raw pointer valid for zero sized reads if the RelVec didn’t allocate.

pub fn as_mut_ptr(this: Mut<'_, Self>) -> *mut T

Returns an unsafe mutable pointer to the RelVec’s buffer, or a dangling raw pointer valid for zero sized reads if the RelVec didn’t allocate.

pub unsafe fn set_len(this: Mut<'_, Self>, new_len: usize)

Forces the length of the vector to new_len.

Safety

new_len must be less than or equal to capacity().

The elements at old_len..new_len must be initialized.

pub fn capacity(&self) -> usize

Returns the maximum number of elements the RelVec can contain before resizing.

pub fn allocator(this: Ref<'_, Self>) -> Ref<'_, A>

Returns a reference to the underlying allocator.

pub fn as_slice(this: Ref<'_, Self>) -> Ref<'_, [T]>

Returns a Ref to a slice of the elements in the RelVec.

pub fn reserve(this: Mut<'_, Self>, additional: usize)where
T: Move<A::Region>,

Reserves capacity for at least additional more elements to be inserted in the given RelVec<T>. The collection may reserve more space to speculatively avoid frequent reallocations. After calling reserve, the capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

pub fn push<E>(this: Mut<'_, Self>, value: E)where
T: Move<A::Region>,
E: Emplace<T, A::Region>,

Appends an element to the back of a collection.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

pub fn extend<I>(this: Mut<'_, Self>, values: I)where
T: Move<A::Region>,
I: Iterator,
I::Item: Emplace<T, A::Region>,

Extends the RelVec with the contents of an iterator.

pub fn clear(this: Mut<'_, Self>)where
T: DropRaw,

Clears the RelVec, removing all values.

Note that this method has no effect on the allocated capacity of the RelVec.